cli train function working
